Extract `validation create` command
